+++ src/love/fonts.lua
@@ -9,7 +9,7 @@ for _,font in pairs{'comic', 'vcr'} do
 		add(font, size)
 	end
 end
-return {function(font, size)
+return function(font, size)
 	if not fonts[font] then fonts[font] = {} end
 	if not fonts[font][size] then 
 		add(font, size)
@@ -17,4 +17,4 @@ return {function(font, size)
 	curFont = fonts[font][size]
 	love.graphics.setFont(fonts[font][size])
 	return fonts[font][size]
-end, fonts['comic'][24]} --default font
+end